+++ buildexe/buildexe.cpp
@@ -17,6 +17,7 @@
 #include "buildcc.h"
 
 #include "buildexe/args_setup.h"
+#include "buildexe/toolchain_setup.h"
 
 #include "bootstrap/build_buildcc.h"
 #include "bootstrap/build_cli11.h"
@@ -41,8 +42,6 @@ static const std::unordered_map<const char *, BuildExeMode> kBuildExeModeMap{
 };
 
 static void clean_cb();
-
-static void host_toolchain_verify(const BaseToolchain &toolchain);
 
 static fs::path get_env_buildcc_home();
 
@@ -116,7 +115,9 @@ int main(int argc, char **argv) {
   }
 
   if (mode == BuildExeMode::Script) {
+    env::log_info(kTag, "*** Starting Toolchain verification ***");
     host_toolchain_verify(toolchain);
+    env::log_info(kTag, "*** Toolchain verification done ***");
   }
 
   PersistentStorage storage;
@@ -197,59 +198,6 @@ int main(int argc, char **argv) {
 static void clean_cb() {
   env::log_info(kTag, fmt::format("Cleaning {}", env::get_project_build_dir()));
   fs::remove_all(env::get_project_build_dir());
-}
-
-static void host_toolchain_verify(const BaseToolchain &toolchain) {
-  env::log_info(kTag, "*** Starting Toolchain verification ***");
-
-  fs::path file = env::get_project_build_dir() / "verify_host_toolchain" /
-                  "verify_host_toolchain.cpp";
-  fs::create_directories(file.parent_path());
-  std::string file_data = R"(// Generated by BuildExe
-#include <filesystem>
-#include <iostream>
-
-namespace fs = std::filesystem;
-
-int main() {
-  std::cout << "********************" << std::endl;
-  std::cout << "Verifying host toolchain" << std::endl;
-  std::cout << "Current Path: " << fs::current_path() << std::endl;
-  std::cout << "********************" << std::endl;
-  return 0;
-})";
-  env::save_file(path_as_string(file).c_str(), file_data, false);
-
-  ExecutableTarget_generic target(
-      "verify", toolchain, TargetEnv(file.parent_path(), file.parent_path()));
-
-  target.AddSource(file);
-  switch (toolchain.GetId()) {
-  case ToolchainId::Gcc:
-  case ToolchainId::MinGW:
-    target.AddCppCompileFlag("-std=c++17");
-    break;
-  case ToolchainId::Msvc:
-    target.AddCppCompileFlag("/std:c++17");
-    break;
-  default:
-    env::assert_fatal<false>("Invalid Compiler Id");
-  }
-  target.Build();
-
-  // Build
-  tf::Executor executor;
-  executor.run(target.GetTaskflow());
-  executor.wait_for_all();
-  env::assert_fatal(env::get_task_state() == env::TaskState::SUCCESS,
-                    "Input toolchain could not compile host program. "
-                    "Requires HOST toolchain");
-
-  // Run
-  bool execute = env::Command::Execute(fmt::format(
-      "{executable}", fmt::arg("executable", target.GetTargetPath().string())));
-  env::assert_fatal(execute, "Could not execute verification target");
-  env::log_info(kTag, "*** Toolchain verification done ***");
 }
 
 static fs::path get_env_buildcc_home() {
